+++ CHANGELOG.md
@@ -27,6 +27,7 @@ jsx-slack v2 has improved JSX structure and built-in components to output the re
 
 - React-compatible public APIs: `JSXSlack.createElement`, `JSXSlack.isValidElement`, and `JSXSlack.Children` helpers
 - HTML-compatible `<Option selected>` and `<RadioButton checked>`
+- `value` prop as an alias into `initialXXX` prop in some interactive components
 - Added JSDoc to many public APIs and components
 
 ### Removed